decimals

+++ WebConversorSolucion/WebConversor/Services/Interfaces/IApiService.cs
@@ -2,6 +2,6 @@
 
 public interface IApiService
 {
-    Task<Dictionary<string, object>> GetDataFromApiAsync(string fromCurrency, string toCurrency, int amount);
+    Task<Dictionary<string, object>> GetDataFromApiAsync(string fromCurrency, string toCurrency, decimal amount);
 
 }
